Fixed Review comments
