+++ composer.json
@@ -92,8 +92,8 @@
         "matthimatiker/opcache-bundle": "^0.7",
         "qandidate/toggle-bundle": "^1.1.0",
         "google/recaptcha": "~1.1",
-        "ggggino/wordbundle": "^1.0",
-        "phpoffice/phpword": "v0.14.*"
+        "phpoffice/phpword": "dev-master",
+        "tecnickcom/tcpdf": "6.*"
     },
     "require-dev": {
         "roave/security-advisories" :"dev-master",
@@ -115,8 +115,7 @@
         "squizlabs/php_codesniffer": "^3.2",
         "escapestudios/symfony2-coding-standard": "^3.1",
         "friendsofphp/php-cs-fixer": "^2.10",
-        "ericmartel/codeception-email-mailtrap": "^1.0",
-        "tecnickcom/tcpdf": "6.*"
+        "ericmartel/codeception-email-mailtrap": "^1.0"
     },
     "scripts": {
         "symfony-scripts": [
